WIP: Add Scylla backend
This adds a first backend for Scylla which simply uses the existing
CQL backend but excludes the DataStax CQL driver and instead uses the
Scylla driver.
It seems to be already working when the library is included directly,
but JanusGraph Server for example doesn't use the Scylla driver yet.

The problem seems to be that both drivers are on the classpath and
the DataStax driver is used because it's found first probably.

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>