Ensure Scylla driver is used for scylla-hadoop
I don't know why this exclusion is necessary because
`cassandra-driver-core` is only an optional dependency of
`janusgraph-cql` and should therefore not be included in the classpath
of `janusgraph-scylla` any way, but it looks like that is not the case.

At least we got failing Scylla Hadoop tests after trying to update the
Scylla driver to 3.11.5. They fail with a `NoSuchMethodError` on
`ResultSet.one()` which was removed in the Scylla driver.

@Bouncheck mentioned that this error should only be possible if we
actually have another version of the driver on our classpath which let
to this (hopefully) fix.

So it should now be possible to update the Scylla driver again.

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>